+++ trader/exchanger/Exchanger.hpp
@@ -2,6 +2,7 @@
 
 #include "core/Proxy.hpp"
 #include "exchanger/base/Storage.hpp"
+#include "exchanger/base/Decimal.hpp"
 #include <cmath>
 
 namespace core {
@@ -39,7 +40,7 @@ public: // methods
     const OrderWrapper* createOrder(OrderRequest& request);
     bool stake(StakingRequest& request);
 
-    double roundQuantity(double quantity, const std::string& key, double(*fn)(double) = std::round) const;
+    Decimal roundQuantity(Decimal quantity, const std::string& key, double(*fn)(double) = std::round) const;
     double fee() const;
 
 protected: // methods
