+++ trader/exchanger/abstract/ExchangerController.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "exchanger/base/Storage.hpp"
+#include "exchanger/base/Decimal.hpp"
 #include <cmath>
 
 namespace core {
@@ -44,7 +45,7 @@ public: // virtual
     virtual bool stake(StakingWrapper& container, StakingRequest& request) = 0;
 
     virtual bool updateStaking(StakingWrapper& container) const = 0;
-    virtual double roundQuantity(double quantity, const std::string& symbol, double(*fn)(double)) const = 0;
+    virtual Decimal roundQuantity(Decimal quantity, const std::string& symbol, double(*fn)(double)) const = 0;
     virtual double fee() const = 0;
 
 protected: // methods
